ntoskrnl.exe: Implement KeWaitForMultipleObjects().
We have plenty of space in which to store a kernel handle inside a
DISPATCHER_HEADER. However, the storage is volatile, and there is no way to
know when to close it. Therefore, we create the handle when
KeWaitForMultipleObjects() is called, and keep it open as long as at least
one thread is waiting on the object.

Signed-off-by: Zebediah Figura <z.figura12@gmail.com>